PlanarTriangulation: move the cache scratch accessors to an internal header
They exist only so the hole-fill plan path in this library can reach buffers of the cache, whose type is private to the .cpp; the public header keeps just the cache handle, its factory and the defaulted cache params.
